fix(finegrained): resolve float precision bug in arithmetic ground truth
- Fix integer overflow in add/sub/mul operations by using native Python
  integer arithmetic instead of float conversion
- Add fix_arithmetic_ground_truth.py script to correct existing results
- Update analysis.py with improved plot layouts (bottom legends,
  transparent backgrounds, per-model bootstrap CIs)
- Add embedding visualization scripts for NL vs code/sim comparisons
- Update generate_plots.py for sim_reasoning vs NL comparisons
- Add algorithm name filtering to prevent label leakage in MI estimation

+++ src/exps_performance/problems/finegrained.py
@@ -81,7 +81,7 @@ class FgCheckAndFormat(CheckAndFormat):
         a = sample_int(d, rng)
         b = sample_int(d, secondary_rng)
         question = f"Compute: {a} + {b}"
-        answer = int(float(a) + float(b))
+        answer = a + b  # Use integer arithmetic to avoid float precision loss
         return FgQuestion(kind=self.k, digits=d, question=question, answer=str(answer))
 
     def load_data(self) -> list[FgQuestion]:
@@ -110,7 +110,7 @@ class SubCheckAndFormat(FgCheckAndFormat):
         question = f"Compute: {a} - {b}"
         if b > a:
             a, b = b, a
-        answer = int(float(a) - float(b))
+        answer = a - b  # Use integer arithmetic to avoid float precision loss
         return FgQuestion(kind=self.k, digits=d, question=question, answer=str(answer))
 
 
@@ -123,7 +123,7 @@ class MulCheckAndFormat(FgCheckAndFormat):
         a = sample_int(d, rng)
         b = sample_int(d, secondary_rng)
         question = f"Compute: {a} * {b}"
-        answer = int(float(a) * float(b))
+        answer = a * b  # Use integer arithmetic to avoid float precision loss
         return FgQuestion(kind=self.k, digits=d, question=question, answer=str(answer))
